+++ NEWS.md
@@ -1,5 +1,7 @@
 # FSA 0.4.26 ongoing
-* `lencat()`: Modified.  Moved all internal functions outside of `lencat()` environment (and renamed them).
+* `lencat()`: Modified.  Fixed bug with category names when `use.names=TRUE`.  Moved all internal functions outside of `lencat()` environment (and renamed them).  Cleaned up code.
+* `psdCalc()`: Modified.  Completely redone.
+* `psdPlot()`: Modified.  Completely redone (fixed several bugs and overall sloppy code).
 
 # FSA 0.4.25 Sep14
 * `mrClosed()`: Modified.  Better handled a given value of `R=`.